refactor: modularize configuration and utilities
