+++ mefrupALS.py
@@ -19,413 +19,9 @@ matplotlib.use("Agg")
 import matplotlib.pyplot as plt
 import plotly.graph_objects as go
 
-# ---------- rutas / const ----------
-BASE_DIR    = os.path.dirname(os.path.abspath(__file__))
-
-# CSVs por máquina
-MACHINES = [
-    {
-        "id": "arburg",
-        "name": "ARBURG 320C GOLDEN EDITION",
-        "oee_csv":  os.path.join(BASE_DIR, "oee_arburg.csv"),
-        "down_csv": os.path.join(BASE_DIR, "down_arburg.csv")
-    },
-    {
-        "id": "yizumi",
-        "name": "YIZUMI UN90 A5",
-        "oee_csv":  os.path.join(BASE_DIR, "oee_yizumi.csv"),
-        "down_csv": os.path.join(BASE_DIR, "down_yizumi.csv")
-    }
-]
-
-# diarios (compat) + combinado
-DAILY_CSV_GLOBAL   = os.path.join(BASE_DIR, "oee_daily.csv")
-DAILY_CSV_INJECTOR = os.path.join(BASE_DIR, "oee_inyeccion_daily.csv")
-
-RECIPES_CSV = os.path.join(BASE_DIR, "recipes.csv")
-LOGO_PATH   = os.path.join(BASE_DIR, "10b41fef-97af-4e79-90c4-b496e0dd3197.png")
-
-CHART_W = 640
-CHART_H = 360
-
-# planificación y milestones
-PLANNING_CSV = os.path.join(BASE_DIR, "planning.csv")   # orden, parte, molde_id, maquina_id, qty_total, inicio_ts, fin_est_ts, setup_min, estado, ciclo_s, cav_on
-DELIV_CSV    = os.path.join(BASE_DIR, "deliveries.csv") # orden, due_date, qty, cumplido(0/1)
-
-# NUEVO: salidas / embarques
-SHIPMENTS_CSV = os.path.join(BASE_DIR, "shipments.csv") # orden, ship_date, qty, destino, nota, approved
-
-OPERADORES   = ["OPERADOR 1", "OPERADOR 2", "OPERADOR 3"]
-TURNOS_HORAS = {1: 8, 2: 8, 3: 8}
-DIAS_ES      = ["Lunes","Martes","Miércoles","Jueves","Viernes","Sábado","Domingo"]
-
-TICK_MS     = 1000
-DEBOUNCE_MS = 160
-DASH_REFRESH_MS = 5000
-
-MOTIVOS_PARO = [
-    "Cambio de molde",
-    "Pieza Atorada",
-    "Sin operador",
-    "Calidad",
-    "Mantenimiento",
-    "Energía (Se fue la luz)"
-]
-
-# ---------- CSV utils ----------
-def asegurar_csv(path, header):
-    try:
-        with open(path, "x", newline="", encoding="utf-8") as f:
-            csv.writer(f).writerow(header)
-    except FileExistsError:
-        pass
-
-def leer_csv_dict(path):
-    if not os.path.exists(path): return []
-    with open(path, "r", newline="", encoding="utf-8") as f:
-        return list(csv.DictReader(f))
-
-def leer_shipments():
-    """Lee shipments.csv garantizando la columna 'approved'."""
-    rows = leer_csv_dict(SHIPMENTS_CSV)
-    for r in rows:
-        if "approved" not in r or r.get("approved") == "":
-            r["approved"] = "1"
-    return rows
-
-def escribir_daily(path, fecha_iso, oee_pct, total, scrap, meta):
-    asegurar_csv(path, ["fecha","oee_dia_%","total_pzs","scrap_pzs","meta_pzs"])
-    rows = leer_csv_dict(path)
-    for r in rows:
-        if r.get("fecha")==fecha_iso:
-            r["oee_dia_%"]=f"{oee_pct:.2f}"; r["total_pzs"]=str(total)
-            r["scrap_pzs"]=str(scrap); r["meta_pzs"]=str(meta); break
-    else:
-        rows.append({"fecha":fecha_iso,"oee_dia_%":f"{oee_pct:.2f}",
-                     "total_pzs":str(total),"scrap_pzs":str(scrap),"meta_pzs":str(meta)})
-    with open(path,"w",newline="",encoding="utf-8") as f:
-        w=csv.DictWriter(f,fieldnames=["fecha","oee_dia_%","total_pzs","scrap_pzs","meta_pzs"])
-        w.writeheader(); w.writerows(rows)
-
-def fechas_registradas(path_daily):
-    asegurar_csv(path_daily, ["fecha","oee_dia_%","total_pzs","scrap_pzs","meta_pzs"])
-    return {r["fecha"] for r in leer_csv_dict(path_daily) if r.get("fecha")}
-
-def asegurar_archivos_basicos():
-    asegurar_csv(DAILY_CSV_GLOBAL, ["fecha","oee_dia_%","total_pzs","scrap_pzs","meta_pzs"])
-    asegurar_csv(DAILY_CSV_INJECTOR, ["fecha","oee_dia_%","total_pzs","scrap_pzs","meta_pzs"])
-    asegurar_csv(RECIPES_CSV, [
-        "molde_id","parte","ciclo_ideal_s","cavidades","cavidades_habilitadas","scrap_esperado_pct","activo"
-    ])
-    if not leer_csv_dict(RECIPES_CSV):
-        with open(RECIPES_CSV,"a",newline="",encoding="utf-8") as f:
-            w=csv.writer(f)
-            w.writerow(["48","19-001-049","45","1","1","2","1"])
-            w.writerow(["84","19-001-084","23","1","1","2","1"])
-    asegurar_csv(PLANNING_CSV, ["orden","parte","molde_id","maquina_id","qty_total","inicio_ts","fin_est_ts","setup_min","estado","ciclo_s","cav_on"])
-    asegurar_csv(DELIV_CSV,    ["orden","due_date","qty","cumplido"])
-    asegurar_csv(SHIPMENTS_CSV,["orden","ship_date","qty","destino","nota","approved"])
-
-def asegurar_archivos_maquina(machine):
-    asegurar_csv(machine["oee_csv"], [
-        "timestamp","fecha","operador","turno","molde","parte","ciclo_s","horas_turno",
-        "tiempo_paro_min","meta_oper_pzs","total_pzs","scrap_pzs","buenas_pzs",
-        "availability_%","performance_%","quality_%","oee_%"
-    ])
-    asegurar_csv(machine["down_csv"], [
-        "fecha","inicio_ts","fin_ts","duracion_seg","motivo","nota","operador","turno","molde"
-    ])
-
-# ---------- cálculos ----------
-def calcular_tiempos(horas_turno, ciclo_s, paro_seg):
-    turno_seg = int(max(0, horas_turno or 0) * 3600)
-    operativo = max(0, turno_seg - int(max(0, paro_seg or 0)))
-    ciclo     = int(ciclo_s or 0)
-    meta_plan = int(turno_seg // ciclo) if ciclo>0 else 0
-    meta_oper = int(operativo // ciclo) if ciclo>0 else 0
-    return turno_seg, operativo, meta_plan, meta_oper
-
-def calcular_metricas(total, scrap, turno_seg, oper_seg, ciclo_ideal_s):
-    total   = int(total or 0)
-    scrap   = int(scrap or 0)
-    buenas  = max(0, total - scrap)
-    A = (oper_seg/turno_seg) if turno_seg>0 else 0.0
-    A = max(0.0, min(1.0, A))
-    perf_num = buenas * float(ciclo_ideal_s or 0)
-    P = (perf_num/oper_seg) if oper_seg>0 else 0.0
-    P = max(0.0, min(1.0, P))
-    Q = (buenas/total) if total>0 else 0.0
-    OEE = A * P * Q * 100.0
-    return buenas, round(A*100,2), round(P*100,2), round(Q*100,2), round(OEE,2)
-
-
-def acum_por_fecha(rows, fecha_iso):
-    total = scrap = 0
-    meta = 0.0
-    n = 0
-    for r in rows:
-        if r.get("fecha") != fecha_iso:
-            continue
-        total += parse_int_str(r.get("total_pzs", "0"))
-        scrap += parse_int_str(r.get("scrap_pzs", "0"))
-        meta += _safe_float(r.get("meta_oper_pzs", "0"))
-        n += 1
-    if total<=0 or meta<=0:
-        return {"count":n,"total":total,"scrap":scrap,"buenas":max(0,total-scrap),
-                "perf_pct":0.0,"qual_pct":0.0,"oee_pct":0.0,"meta_pzs":meta}
-    buenas=max(0,total-scrap); P=total/meta; Q=buenas/total; OEE=P*Q*100.0
-    return {"count":n,"total":total,"scrap":scrap,"buenas":buenas,
-            "perf_pct":round(P*100,2),"qual_pct":round(Q*100,2),
-            "oee_pct":round(OEE,2),"meta_pzs":meta}
-
-def acum_global(rows):
-    total = scrap = 0
-    meta = 0.0
-    n = 0
-    dias = set()
-    for r in rows:
-        total += parse_int_str(r.get("total_pzs", "0"))
-        scrap += parse_int_str(r.get("scrap_pzs", "0"))
-        meta += _safe_float(r.get("meta_oper_pzs", "0"))
-        n += 1
-        if r.get("fecha"):
-            dias.add(r["fecha"])
-    if total<=0 or meta<=0:
-        return {"registros":n,"dias":len(dias),"total":total,"scrap":scrap,
-                "buenas":max(0,total-scrap),"perf_pct":0.0,"qual_pct":0.0,"oee_pct":0.0,"meta_pzs":meta}
-    buenas=max(0,total-scrap); P=total/meta; Q=buenas/total; OEE=P*Q*100.0
-    return {"registros":n,"dias":len(dias),"total":total,"scrap":scrap,"buenas":buenas,
-            "perf_pct":round(P*100,2),"qual_pct":round(Q*100,2),"oee_pct":round(OEE,2),"meta_pzs":meta}
-
-def promedio_oee_daily(path_daily):
-    rows = leer_csv_dict(path_daily) if os.path.exists(path_daily) else []
-    vals = [_safe_float(r.get("oee_dia_%", 0)) for r in rows]
-    return round(sum(vals)/len(vals),2) if vals else 0.0
-
-
-def resumen_historico_maquina(machine):
-    """Promedios históricos de OEE y sus componentes para una máquina."""
-    asegurar_archivos_maquina(machine)
-    rows = leer_csv_dict(machine["oee_csv"])
-    if not rows:
-        return dict(oee=0.0, A=0.0, P=0.0, Q=0.0)
-
-    def _avg(key):
-        vals = [_safe_float(r.get(key, 0)) for r in rows if r.get(key) not in (None, "")]
-        return sum(vals) / len(vals) if vals else 0.0
-
-    return dict(
-        oee=round(_avg("oee_%"), 2),
-        A=round(_avg("availability_%"), 2),
-        P=round(_avg("performance_%"), 2),
-        Q=round(_avg("quality_%"), 2),
-    )
-
-
-def dia_semana_es(f):
-    try: y,m,d=map(int,f.split("-")); return DIAS_ES[date(y,m,d).weekday()]
-    except: return "Día"
-
-def segs_to_hms_str(s):
-    s=max(0,int(s)); h=s//3600; m=(s%3600)//60; sc=s%60; return f"{h:02d}:{m:02d}:{sc:02d}"
-
-def parse_int_str(s, default=0):
-    try:
-        s = str(s).replace(",",".").strip()
-        return int(float(s))
-    except:
-        return default
-
-# ===== producidas por molde (para milestones/órdenes) =====
-def producido_por_molde_global(molde_id: str, hasta_fecha: str = None) -> int:
-    total_buenas = 0
-    for m in MACHINES:
-        rows = leer_csv_dict(m["oee_csv"])
-        for r in rows:
-            try:
-                if str(r.get("molde","")).strip() == str(molde_id).strip():
-                    if hasta_fecha:
-                        if (r.get("fecha") or "") <= hasta_fecha:
-                            total_buenas += int(float(r.get("buenas_pzs","0")))
-                    else:
-                        total_buenas += int(float(r.get("buenas_pzs","0")))
-            except: pass
-    return total_buenas
-
-def enviados_por_orden(orden: str) -> int:
-    return sum(
-        parse_int_str(r.get("qty", "0"))
-        for r in leer_shipments()
-        if r.get("orden") == orden and r.get("approved", "0") == "1"
-    )
-
-def enviados_por_molde(molde_id: str) -> int:
-    """Cantidad total enviada asociada a un molde (todas las órdenes)."""
-    orden_a_molde = {
-        r.get("orden", "").strip(): r.get("molde_id", "").strip()
-        for r in leer_csv_dict(PLANNING_CSV)
-    }
-    total = 0
-    for r in leer_shipments():
-        if (
-            orden_a_molde.get(r.get("orden", "").strip()) == str(molde_id).strip()
-            and r.get("approved", "0") == "1"
-        ):
-            total += parse_int_str(r.get("qty", "0"))
-    return total
-
-# ===== resumen por máquina (hoy) con fallback robusto =====
-def _safe_float(x, default=0.0):
-    """Parsea un valor numérico tolerante removiendo comas y signos %."""
-    try:
-        if x in (None, ""):
-            return default
-        s = str(x).replace(",", ".").replace("%", "").strip()
-        return float(s)
-    except Exception:
-        return default
-
-def resumen_hoy_maquina(machine, fecha_iso):
-    asegurar_archivos_maquina(machine)
-    rows = [r for r in leer_csv_dict(machine["oee_csv"]) if r.get("fecha")==fecha_iso]
-    if not rows:
-        return dict(oee=0.0,A=0.0,P=0.0,Q=0.0,total=0,buenas=0,scrap=0,
-                    meta=0, ciclo_ideal=0, ciclo_real=0.0,
-                    turno_seg=0, oper_seg=0, ultimo_paro="-")
-    # acumulados
-    total = sum(parse_int_str(r.get("total_pzs","0")) for r in rows)
-    scrap = sum(parse_int_str(r.get("scrap_pzs","0")) for r in rows)
-    buenas = max(0, total - scrap)
-    # meta planeada almacenada
-    meta_oper = sum(parse_int_str(r.get("meta_oper_pzs", "0")) for r in rows)
-    # tiempos
-    horas_sum = sum(_safe_float(r.get("horas_turno", "0")) for r in rows)
-    turno_seg = int(horas_sum * 3600)
-    paro_seg = int(sum(_safe_float(r.get("tiempo_paro_min", "0")) for r in rows) * 60)
-    oper_seg = max(0, turno_seg - paro_seg)
-    # ciclo ideal (último válido)
-    ciclos = [parse_int_str(r.get("ciclo_s", "0")) for r in rows if parse_int_str(r.get("ciclo_s", "0")) > 0]
-    ciclo_ideal = ciclos[-1] if ciclos else 0
-
-    A = (oper_seg / turno_seg) if turno_seg > 0 else None
-    P = (total / meta_oper) if meta_oper > 0 else None
-    Q = (buenas / total) if total > 0 else None
-    if None in (A, P, Q):
-        # fallback a promedios almacenados si faltan datos para calcular
-        n = max(1, len(rows))
-        A = sum(_safe_float(r.get("availability_%", "0")) for r in rows) / n / 100.0
-        P = sum(_safe_float(r.get("performance_%", "0")) for r in rows) / n / 100.0
-        Q = sum(_safe_float(r.get("quality_%", "0")) for r in rows) / n / 100.0
-    oee = A * P * Q * 100.0
-    A *= 100.0; P *= 100.0; Q *= 100.0
-
-    # ciclo real estimado
-    ciclo_real = (oper_seg / buenas) if buenas > 0 else 0.0
-    # último paro
-    downs = [r for r in leer_csv_dict(machine["down_csv"]) if r.get("fecha")==fecha_iso]
-    if downs:
-        d = downs[-1]
-        try:
-            mins = round(int(float(d.get("duracion_seg","0")))/60.0,1)
-            ultimo = f"{d.get('inicio_ts','')} -> {d.get('fin_ts','')}   {mins:.1f} min ({d.get('motivo','')})"
-        except:
-            ultimo = f"{d.get('inicio_ts','')} -> {d.get('fin_ts','')}   ({d.get('motivo','')})"
-    else:
-        ultimo="-"
-    return dict(
-        oee=round(oee, 2),
-        A=round(A, 2),
-        P=round(P, 2),
-        Q=round(Q, 2),
-        total=total,
-        buenas=buenas,
-        scrap=scrap,
-        meta=int(meta_oper),
-        ciclo_ideal=ciclo_ideal,
-        ciclo_real=round(ciclo_real, 2),
-        turno_seg=turno_seg,
-        oper_seg=oper_seg,
-        ultimo_paro=ultimo,
-    )
-
-def resumen_rango_maquina(machine, desde, hasta):
-    """Agrega métricas de producción en un rango de fechas."""
-    asegurar_archivos_maquina(machine)
-    rows = leer_csv_dict(machine["oee_csv"])
-    down_rows = leer_csv_dict(machine["down_csv"])
-    down_by_date = {}
-    for d in down_rows:
-        f = d.get("fecha")
-        if not f:
-            continue
-        if desde and f < desde:
-            continue
-        if hasta and f > hasta:
-            continue
-        mins = _safe_float(d.get("duracion_seg", "0")) / 60.0
-        down_by_date[f] = down_by_date.get(f, 0.0) + mins
-
-    data = []
-    for r in rows:
-        f = r.get("fecha")
-        if not f:
-            continue
-        if desde and f < desde:
-            continue
-        if hasta and f > hasta:
-            continue
-        total = parse_int_str(r.get("total_pzs", "0"))
-        scrap = parse_int_str(r.get("scrap_pzs", "0"))
-        meta = _safe_float(r.get("meta_oper_pzs", "0"))
-        avail = _safe_float(r.get("availability_%", "0"))
-        perf = _safe_float(r.get("performance_%", "0"))
-        qual = _safe_float(r.get("quality_%", "0"))
-        oee = _safe_float(r.get("oee_%", "0"))
-        if total <= 0 or meta <= 0:
-            continue
-        buenas = max(0, total - scrap)
-        paro = round(down_by_date.get(f, 0.0), 2)
-        data.append({
-            "fecha": f,
-            "paro_min": paro,
-            "total": total,
-            "scrap": scrap,
-            "buenas": buenas,
-            "availability": round(avail, 2),
-            "performance": round(perf, 2),
-            "quality": round(qual, 2),
-            "oee": round(oee, 2)
-        })
-    if not data:
-        return {
-            "total": 0,
-            "scrap": 0,
-            "buenas": 0,
-            "oee_prom": 0.0,
-            "avail_prom": 0.0,
-            "perf_prom": 0.0,
-            "qual_prom": 0.0,
-            "paro_min_total": 0.0,
-        }, []
-    total = sum(d["total"] for d in data)
-    scrap = sum(d["scrap"] for d in data)
-    buenas = sum(d["buenas"] for d in data)
-    oee_prom = sum(d["oee"] for d in data) / len(data)
-    avail_prom = sum(d["availability"] for d in data) / len(data)
-    perf_prom = sum(d["performance"] for d in data) / len(data)
-    qual_prom = sum(d["quality"] for d in data) / len(data)
-    total_paro = sum(d["paro_min"] for d in data)
-    stats = {
-        "total": total,
-        "scrap": scrap,
-        "buenas": buenas,
-        "oee_prom": round(oee_prom, 2),
-        "avail_prom": round(avail_prom, 2),
-        "perf_prom": round(perf_prom, 2),
-        "qual_prom": round(qual_prom, 2),
-        "paro_min_total": round(total_paro, 2),
-    }
-    return stats, data
+from config import *
+from csv_utils import *
+from metrics import *
 
 # ---------- Vistas ----------
 class RecipesView(ctk.CTkFrame):
